requirements.txt
langchain
langchain-openai
pydantic-ai[openai]
pydantic-ai-slim[duckduckgo]
langgraph
crewai
python-dotenv
requests
openai
pytest